Refactor -> organización de directorios MVP
